For #16477: Migrate updateAccessibilityCollectionItemInfo to extension.
